🚚 Move navigation to patterns
