+++ src/patterns/index.ts
@@ -1,1 +1,2 @@
 export * from "./Footer/Footer"
+export * from "./Navigation/Navigation"
